# ssb-app-server/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Spice Spice Baby Application Server",
    "description": "SSB Recipe App back-end application server",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Recipe database API endpoints"
    }
  ],
...
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/recipes/{recipeId}": {
      "parameters": [
        {
          "name": "recipeId",
          "in": "path",
          "required": true,
          "description": "Recipe ID",
          "type": "string"
        }
      ],
      "get": {
        "tags": ["Recipes"],
        "summary": "Get recipe with given Recipe ID",
        "parameters": [
          {
            "name": "recipeId",
            "in": "path",
            "required": true,
            "description": "Recipe ID",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Recipe"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Recipe": {
      "type": "object",
      "properties": {
        "recipe_id": { "type": "IIdentity" },
        "recipe_name": { "type": "String" },
        "recipe_picture": { "type": "String" },
        "all_ingredients": { "type": "String[]" },
        "main_ingredients": { "type": "String[]" },
        "ingredient_amounts": { "type": "String[]" },
        "ingredient_units": { "type": "String[]" },
        "ingredient_prep": { "type": "String[]" }
      }
    },
...
  }
}